+++ .eslintignore
@@ -12,3 +12,4 @@ src/utils.ts
 src/@types/
 script/lib/
 script/build.ts
+config/types/
