+++ config/filters/isColorWithAlpha.test.ts
@@ -1,4 +1,4 @@
-import {getMockToken} from '../../test-utilities/getMockToken'
+import {getMockToken} from '../../src/test-utilities/getMockToken'
 import {isColorWithAlpha} from './isColorWithAlpha'
 
 describe('Filter: isColorWithAlpha', () => {
