+++ config/parsers/w3c-json-parser.test.ts
@@ -1,5 +1,5 @@
 import {w3cJsonParser} from './w3c-json-parser'
-import {output as expectedOutput, getMockParserInput} from '../../test-utilities/getMockParserInput'
+import {output as expectedOutput, getMockParserInput} from '../../src/test-utilities/getMockParserInput'
 
 describe('Parser: w3c token json5 parser', () => {
   it('parses valid w3c json5', () => {
