+++ config/transformers/color-to-hex-alpha.test.ts
@@ -1,4 +1,4 @@
-import {getMockToken} from '../../test-utilities/getMockToken'
+import {getMockToken} from '../../src/test-utilities/getMockToken'
 import {colorToHexAlpha} from './color-to-hex-alpha'
 
 describe('Transformer: colorToHexAlpha', () => {
