+++ config/transformers/color-to-rgb-alpha.test.ts
@@ -1,4 +1,4 @@
-import {getMockToken} from '../../test-utilities/getMockToken'
+import {getMockToken} from '../../src/test-utilities/getMockToken'
 import {colorToRgbAlpha} from './color-to-rgb-alpha'
 
 describe('Transformer: colorToRgbAlpha', () => {
